+++ server/src/tests/mod.rs
@@ -5,3 +5,4 @@ pub mod loc_version;
 pub mod parser_skip;
 pub mod scripted_loc;
 pub mod utf16_conversion;
+pub mod formatting;